refactor: Phase 1 基础设施升级 — Java 21 + Spring Boot 3.3.5 + jakarta
- Java 8→21, Spring Boot 2.6.7→3.3.5
- javax.* → jakarta.* (servlet/validation/annotation, ~80 files)
- MyBatis-Plus 3.4.0→3.5.9, Netty 4.1.76→4.1.114, 等 10+ 依赖升级
- Springfox/Swagger → 空桩过渡 (Phase 2 换 Springdoc)
- 配置迁移: 移除 pathmatch, 修正 logback 日期格式
- 全 8 模块编译+打包通过

+++ mallchat-chat-server/src/main/java/com/abin/mallchat/common/chat/dao/ContactDao.java
@@ -39,20 +39,20 @@ public class ContactDao extends ServiceImpl<ContactMapper, Contact> {
                 .eq(Contact::getRoomId, message.getRoomId())
                 .ne(Contact::getUid, message.getFromUid())// 不需要查询出自己
                 .ge(Contact::getReadTime, message.getCreateTime())
-                .count();
+                .count().intValue();
     }
 
     public Integer getTotalCount(Long roomId) {
         return lambdaQuery()
                 .eq(Contact::getRoomId, roomId)
-                .count();
+                .count().intValue();
     }
 
     public Integer getUnReadCount(Message message) {
         return lambdaQuery()
                 .eq(Contact::getRoomId, message.getRoomId())
                 .lt(Contact::getReadTime, message.getCreateTime())
-                .count();
+                .count().intValue();
     }
 
     public CursorPageBaseResp<Contact> getReadPage(Message message, CursorPageBaseReq cursorPageBaseReq) {
